remove local discovery and use a much simpler NoneDiscovery for tribes
